# swagger.yaml
swagger: '2.0'
info:
  description: Набор сервисов для работы с системой учета /ООО Ромашка/
  version: "1.0.0"
  title: RestAPI
consumes:
  - application/json
  - application/xml
  - application/text
produces:
  - application/json
  - application/xml
  - application/text
basePath: "/api"

paths:
  /reports/formats:
    get:
      summary: Получить список форматов
      operationId: main.formats
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /reports/{category}/{format}:
    get:
      summary: Получить отчёт по заданным параметрам
      operationId: main.get_report
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка
    parameters:
    - name: category
      in: path
      description: Наименование категории данных (nomenclature, nomenclature_group, measurement, recipe)
      required: true
      type: string
    - name: format
      in: path
      description: Наименование формата отчёта (CSV, MARKDOWN, JSON, XML, RTF)
      required: true
      type: string

  /filter/{domain}:
    post:
      summary: Получить данные по параметрам фильтрации
      operationId: main.filter_data
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка
      parameters:
        - name: domain
          in: path
          required: true
          type: string
          description: Наименование категории данных (nomenclature, nomenclature_group, measurement, recipe)
        - in: body
          name: filter
          description: Параметры фильтрации
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              unique_code:
                type: string
              type:
                type: string

  /transactions:
    post:
      summary: Получить данные о транзакциях
      operationId: main.transactions
      consumes:
        - application/json
      produces:
        - text/html
      parameters:
        - in: body
          name: filter_data
          description: Параметры фильтрации
          required: true
          schema:
            type: object
            properties:
              start_period:
                type: string
              end_period:
                type: string
              warehouse:
                type: object
                properties:
                  name:
                    type: string
                  unique_code:
                    type: string
                  type:
                    type: string
                    enum: ['equals', 'like']
              nomenclature:
                type: object
                properties:
                  name:
                    type: string
                  unique_code:
                    type: string
                  type:
                    type: string
                    enum: ['equals', 'like']
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка

  /turnovers:
    post:
      summary: Получить данные об оборотах
      operationId: main.turnovers
      consumes:
        - application/json
      produces:
        - text/html
      parameters:
        - in: body
          name: filter_data
          description: Параметры фильтрации
          required: true
          schema:
            type: object
            properties:
              start_period:
                type: string
              end_period:
                type: string
              warehouse:
                type: object
                properties:
                  name:
                    type: string
                  unique_code:
                    type: string
                  type:
                    type: string
                    enum: ['equals', 'like']
              nomenclature:
                type: object
                properties:
                  name:
                    type: string
                  unique_code:
                    type: string
                  type:
                    type: string
                    enum: ['equals', 'like']
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка

  /block_period:
    get:
      summary: Получить дату загрузки
      operationId: main.block_period
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /new_block_period:
    post:
      summary: Обновить дату блокировки
      operationId: main.new_block_period
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка
      parameters:
        - in: body
          name: new_block_period
          description: Новая дата блокировки
          required: true
          schema:
            type: object
            properties:
              block_period:
                type: string

  /api/nomenclature/get:
    get:
      summary: Получить номенклатуру
      produces:
        - text/html
      operationId: main.get_nomenclature
      parameters:
        - name: unique_code
          in: query
          description: Уникальный код номенклатуры
          required: true
          type: string
      responses:
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка

  /api/nomenclature/add:
    put:
      summary: Добавить номенклатуру
      produces:
        - text/html
      operationId: main.add_nomenclature
      parameters:
        - in: body
          name: body
          description: Новая номенклатура
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              full_name:
                type: string
              nomenclature_group_unique_code:
                type: string
              measurement_unique_code:
                type: string
      responses:
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка

  /api/nomenclature/update:
    patch:
      summary: Обновить номенклатуру
      produces:
        - text/html
      operationId: main.update_nomenclature
      parameters:
        - in: body
          name: body
          description: Новые данные номенклатуры
          required: true
          schema:
            type: object
            properties:
              unique_code:
                type: string
              name:
                type: string
              full_name:
                type: string
              nomenclature_group_unique_code:
                type: string
              measurement_unique_code:
                type: string
      responses:
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка

  /api/nomenclature/delete:
    delete:
      summary: Удалить номенклатуру
      produces:
        - text/html
      operationId: main.delete_nomenclature
      parameters:
        - in: body
          name: body
          description: Уникальный код номенклатуры
          required: true
          schema:
            type: object
            properties:
              unique_code:
                type: string
      responses:
        500:
          description: Ошибка на сервере
        404:
          description: Данные не найдены
        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка